+++ src/languages/ruby.js
@@ -188,7 +188,10 @@ export default function(hljs) {
           begin: '<\\s*',
           contains: [
             {
-              begin: '(' + hljs.IDENT_RE + '::)?' + hljs.IDENT_RE
+              begin: '(' + hljs.IDENT_RE + '::)?' + hljs.IDENT_RE,
+              // we already get points for <, we don't need poitns
+              // for the name also
+              relevance: 0
             }
           ]
         }
@@ -200,6 +203,7 @@ export default function(hljs) {
       // def method_name;
       // def method_name (end of line)
       begin: regex.concat(/def\s*/, regex.lookahead(RUBY_METHOD_RE + "\\s*(\\(|;|$)")),
+      relevance: 0, // relevance comes from kewords
       keywords: "def",
       end: '$|;',
       contains: [