+++ CHANGES.md
@@ -18,6 +18,7 @@ Grammar changes:
 
 Parser:
 
+- `beginKeyword` no longer bestows double relevance (#2953) [Josh Goebel][]
 - allow `keywords` to be an array of strings [Josh Goebel][]
 - add `modes.MATCH_NOTHING_RE` that will never match
   - This can be used with `end` to hold a mode open (it must then be ended with